# [PERF] product: speedup name_search with positive operators