@robodoo 2025-09-30
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/229275.png)](https://mergebot.odoo.com/odoo/odoo/pull/229275)