@Feyensv 2025-10-07
Honestly not convinced we should make this logic more complex.
If you find some clean quick win I don't mind, but I don't think we should go too far, particularly in stable.

Regarding the numbers you mentioned, as far as I remember, having millions of product goes above what has always been considered/promised as the 'supported' numbers of products odoo is guaranteed to handle in reasonable times.