@chapital 2025-11-08
In odoo 18 in a database of 300K products, when we search a product it take less than a second, but in odoo 19 take 3-5 seconds, its normal?