@LGx0 2025-11-20
Hello,
How can I test this on the current system I’m running on Odoo 19 via Odoo.sh? I currently have 5.5 million products in my database. The back-end at the database level is fast, but the UI is terrible — it takes around 40 seconds just to find a default code ra. I cannot delete a product… I have 5 workers, but I was told by Odoo support that performance is not the issue. I’m using Odoo 19. Could you help me understand how to use this? Thanks.

If you like, I can also polish a version for formal partner/vendor support.

Video: https://drive.google.com/file/d/1b9FciniMy9vqRPMci9hm9sEtz3g0eOOB/view

Could you help me? Thank you.