@jjscarafia 2025-11-24
Databases with only 300 k products (is not a lot) are not usable. Searching products on a sale order lines takes 12 seconds for each new letter (replicated on odoo.sh). Check this video.

The default search on "sale order lines" (purchase, moves, etc) is by product.template. There, the time is about 15 seconds. If you search by variant (product.product) it goes down to 2.5 seconds. (in 16 both times were the same).

Video on 18: https://drive.google.com/file/d/1Uz4DKTAQ5AtpDLEMV76ojEx4IzUdoXR5/view
Video on 16: https://drive.google.com/file/d/1baDBFiQ3N9cQvFLdzBiwr3pnC27kh7_W/view

More info on OPW 5221753

Thanks for working on a fix!